Initial use of clap
