+++ Cargo.toml
@@ -11,7 +11,7 @@ edition = "2021"
 cargo_metadata = "0.14.0"
 rustc_version = "0.4.0"
 semver = "1.0.10"
-serde = "1.0.111"
-serde_derive = "1.0.111"
+serde = {version = "1.0.139", features = ['derive']}
 tee = "0.1.0"
 toml = "0.5.6"
+clap = { version = "3.2.12", features = ["derive"]}
